requirements.txt
pyyaml
httpx 
sqlalchemy 
aiosqlite 
uvicorn 
fastapi 
pytest 
rich 
python-dotenv
opensearch-py
google-generativeai>=0.3.0